% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pause Site Detection.R
\name{get_pause_sites}
\alias{get_pause_sites}
\title{Create a RiboBayes object from a Ribo object}
\usage{
get_pause_sites(ribo, transcripts, p_shift, experiments, cores)
}
\arguments{
\item{ribo}{Ribo object}

\item{transcripts}{Character vector of transcript names to detect pause sites on.}

\item{p_shift}{if \code{FALSE}, P-site adjustments are not applied. Defaults to \code{TRUE}.}

\item{experiments}{Character vector of ribosome profiling experiments to draw data from.}

\item{cores}{Number of cores to run \code{get_pause_sites()} on.}
}
\value{
A RiboBayes object containing []
}
\description{
Takes as input a Ribo object and returns a RiboBayes object containing information on the
location and differential expression data of all ribosome pause sites on the inputted transcripts.
}
