% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream Analysis.R
\name{get_ps_info}
\alias{get_ps_info}
\title{Create a data frame summarizing the output of \code{get_pause_sites()}}
\usage{
get_ps_info(sites)
}
\arguments{
\item{sites}{List of ribosome pause sites; usually the output of \code{get_pause_sites()}}
}
\value{
Data frame with one row for each pause site and columns containing a unique identifier,
transcript location, nucleotide position, and significance score in all replicates for each pause site.
}
\description{
Takes as input a list of ribosome pause sites (the output of \code{get_pause_sites()}) and returns a
data frame with one row for each pause site, with columns containing an unique identifier, transcript location,
nucleotide position, and significance score in all replicates for each pause site.
}
