% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream Analysis.R
\name{plot_dps_per_transcript}
\alias{plot_dps_per_transcript}
\title{Plot bar plots to visualize the distribution of differential pause sites per transcript}
\usage{
plot_dps_per_transcript(sites)
}
\arguments{
\item{sites}{List of ribosome pause sites; usually the output of \code{get_pause_sites()}}
}
\value{
\code{NULL}
}
\description{
Takes as input a list of ribosome pause sites (the output of \code{get_pause_sites()}) and plots
two bar plots: one showing the distribution of the number of differential pause sites per transcript,
and another showing the distribution of the number of upregulated vs. downregulated pause sites per
transcript.
}
