% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream Analysis.R
\name{plot_ps_conservation}
\alias{plot_ps_conservation}
\title{Create Venn Diagrams to visualize the conservation of pause sites across experiments}
\usage{
plot_ps_conservation(sites)
}
\arguments{
\item{sites}{List of ribosome pause sites; usually the output of \code{get_pause_sites()}}

\item{bayesian_adjust}{If \code{TRUE}, shares variance across pause sites using techniques from the edgeR package to determine significance.
If \code{FALSE}, uses a simple two-sample t-test to determine significance. Defaults to \code{TRUE}.}
}
\value{
One figure with two Venn diagrams (one for each condition) depicting the overlap of pause site
expression across experiments. One UpSetR graph providing more detailed information about the expression
of pause sites across replicates and experiments.
}
\description{
Takes as input a list of ribosome pause sites (the output of \code{get_pause_sites()}) and plots
Venn diagrams depicting the overlap of detected pause sites across experiments.
}
