% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream Analysis.R
\name{plot_ps_per_transcript}
\alias{plot_ps_per_transcript}
\title{Create bar plots to visualize the distribution of pause sites per transcript}
\usage{
plot_ps_per_transcript(sites)
}
\arguments{
\item{sites}{List of ribosome pause sites; usually the output of \code{get_pause_sites()}}
}
\value{
\code{NULL}
}
\description{
Takes as input a list of ribosome pause sites (the output of \code{get_pause_sites()}) and plots
a bar plot showing the distribution of pause sites, both constant and differential, per transcript.
}
