% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream Analysis.R
\name{plot_ps_regulation}
\alias{plot_ps_regulation}
\title{Plot violin plots to visualize the distribution of t-statistics for pause site significance.}
\usage{
plot_ps_regulation(sites, bayesian_adjust = TRUE)
}
\arguments{
\item{sites}{List of ribosome pause sites; usually the output of \code{get_pause_sites()}}

\item{bayesian_adjust}{If \code{TRUE}, shares variance across pause sites using techniques from the edgeR package to determine significance.
If \code{FALSE}, uses a simple two-sample t-test to determine significance. Defaults to \code{TRUE}.}
}
\value{
A \code{ggstatsplot} figure with two violin plots: one showing the distribution of pause site
t-statistics for constant pause sites only (p>0.05) and one showing the distribution of pause
site t-statistics of differential pause sites only (p<=0.05).
}
\description{
Takes as input a list of ribosome pause sites (the output of \code{get_pause_sites()}) and plots
two violin plots showing the distribution of t-statistics for constant and differential pause sites.
}
