% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Downstream Analysis.R
\name{plot_ps_volcano}
\alias{plot_ps_volcano}
\title{Plot a volcano plot of -log(p-value) vs. log fold change for all pause sites.}
\usage{
plot_ps_volcano(sites, bayesian_adjust = TRUE)
}
\arguments{
\item{sites}{List of ribosome pause sites; usually the output of \code{get_pause_sites()}}

\item{bayesian_adjust}{If \code{TRUE}, shares variance across pause sites using techniques from the edgeR package to determine significance.
If \code{FALSE}, uses a simple two-sample t-test to determine significance. Defaults to \code{TRUE}.}
}
\value{
\code{NULL}
}
\description{
Takes as input a list of ribosome pause sites (the output of \code{get_pause_sites()}) and
plots a volcano plot showing the log fold change vs. -log(p-value) of change across conditions
for all detected ribosome pause sites. Labels the most significant pause sites (those with highest
log fold change and/or -log(p)).
}
