% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pause Site Detection.R
\name{get_pause_sites}
\alias{get_pause_sites}
\title{Get a list of all pause sites on inputted transcripts with differential expression scoring
from a .ribo object}
\usage{
get_pause_sites(ribo, transcript, shifts, lengths, experiments)
}
\arguments{
\item{ribo}{.ribo file}

\item{transcript}{List of transcripts to find pause sites on}

\item{shifts}{List of P-site shifts (may phase this out)}

\item{lengths}{List of read lengths to use in pause site detection}

\item{experiments}{List of ribosome profiling experiment names in the .ribo object to draw data from}
}
\value{
A RiboBayes object
}
\description{
This function takes as input a .ribo object (and other optional parameters that can be specified)
and returns a RiboBayes object containing information on the location and differential expression data
of all ribosome pause sites on the inputted transcripts.
}
